Add RuntimeStats for validation file count and file write count
