+++ lib/src/dartdoc.dart
@@ -16,9 +16,9 @@ import 'package:dartdoc/src/generator/generator.dart';
 import 'package:dartdoc/src/generator/html_generator.dart';
 import 'package:dartdoc/src/generator/markdown_generator.dart';
 import 'package:dartdoc/src/logging.dart';
-import 'package:dartdoc/src/matching_link_result.dart';
 import 'package:dartdoc/src/model/model.dart';
 import 'package:dartdoc/src/package_meta.dart';
+import 'package:dartdoc/src/runtime_stats.dart';
 import 'package:dartdoc/src/utils.dart';
 import 'package:dartdoc/src/validator.dart';
 import 'package:dartdoc/src/version.dart';
