+++ lib/src/validator.dart
@@ -10,6 +10,7 @@ import 'package:dartdoc/src/dartdoc_options.dart';
 import 'package:dartdoc/src/logging.dart';
 import 'package:dartdoc/src/model/model_element.dart';
 import 'package:dartdoc/src/model/package_graph.dart';
+import 'package:dartdoc/src/runtime_stats.dart';
 import 'package:dartdoc/src/tuple.dart';
 import 'package:dartdoc/src/warnings.dart';
 import 'package:html/parser.dart' show parse;
@@ -35,6 +36,8 @@ class Validator {
   /// generated all docs for the Package.
   void validateLinks() {
     logInfo('Validating...');
+    runtimeStats.accumulators['readCountForLinkValidation'] = 0;
+    runtimeStats.accumulators['readCountForIndexValidation'] = 0;
     _collectLinks('index.html');
     _checkForOrphans();
     _checkSearchIndex();
@@ -144,6 +147,8 @@ class Validator {
       return;
     }
     final jsonData = json.decode(file.readAsStringSync()) as List;
+    runtimeStats.accumulators
+        .update('readCountForIndexValidation', (c) => c + 1);
 
     final found = <String>{};
     found.add(fullPath);
@@ -167,6 +172,9 @@ class Validator {
     }
   }
 
+  /// Gets all link destinations, and the base link destination, if one is found
+  /// on the page.
+  ///
   /// This is extracted to save memory during the check; be careful not to hang
   /// on to anything referencing the full file and doc tree.
   Tuple2<Iterable<String>, String?>? _getStringLinksAndHref(String fullPath) {
@@ -178,6 +186,8 @@ class Validator {
     // `lowercaseElementName: false` and `lowercaseAttrName: false` may save
     // time or memory.
     final doc = parse(file.readAsBytesSync());
+    runtimeStats.accumulators
+        .update('readCountForLinkValidation', (c) => c + 1);
     final base = doc.querySelector('base');
     String? baseHref;
     if (base != null) {
